requirements.txt
# Core Django dependencies
Django==5.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1

# Database
dj-database-url==2.1.0
psycopg[binary]==3.2.9

# Production server
gunicorn==21.2.0

# Static files and CORS
whitenoise==6.6.0
django-cors-headers==4.3.1

# Environment variables
python-dotenv==1.0.1

# Google OAuth
google-auth==2.29.0
requests==2.32.4

# Essential Python dependencies (auto-installed but pinned for stability)
asgiref==3.8.1
sqlparse==0.5.3
pytz==2025.2
